# Skip regionprops construction