Previously we were constructing "region properties" objects, even though we were barely using them. This incurred a very large number of albeit very small allocations & function calls but it added up.

This improves performance by:

- 260k pixel image: 202ms --> 170ms, 32ms 16%

BEFORE:
<img width="1089" alt="Screenshot 2024-10-03 at 1 34 12 PM" src="https://github.com/user-attachments/assets/75d245b2-cc81-4aba-84d3-15a4182e6fd4">

AFTER:
<img width="1088" alt="Screenshot 2024-10-03 at 1 34 38 PM" src="https://github.com/user-attachments/assets/aec138da-9de8-49a5-8bc4-19644113a6e1">


- 140M pixel image: 15.6s --> 11.7s, 3.9s or 25%

just for fill-holes, not the overall which was 113s --> 111s   (recall that these profilings are imprecise using cProfile)

BEFORE: (just for fill_holes)
<img width="1542" alt="Screenshot 2024-10-03 at 1 35 18 PM" src="https://github.com/user-attachments/assets/b53e199b-b96c-4761-b60d-edb557a54013">

AFTER: (just for fill_holes)
<img width="1545" alt="Screenshot 2024-10-03 at 1 36 11 PM" src="https://github.com/user-attachments/assets/ec552b88-d9f5-440c-958e-f2becc70789a">


Fixes #357

Paired with @WeihaoGe1009